requirements.txt
#安装扩展依赖
flask
flask-sqlalchemy
mysqlclient
flask_script
